Add product infomation

+++ src/main/java/co/istad/productapi/dto/product/request/UpdateProductRequest.java
@@ -2,13 +2,15 @@ package co.istad.productapi.dto.product.request;
 
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 public record UpdateProductRequest(
 
         String name,
         String description,
 
         @Positive
-        Float price
+        BigDecimal price
 
 ) {
 }
